rtl8723ds: remove raw_spinlock on RT
There's no need for a raw spinlock in this driver as it doesn't do
anything in hardirq context on -rt.  Some of the critical sections
guarded by these locks are long and contain unbounded loops so there is
a risk of raw spinlocks causing latency spikes on -rt.

The existing code is broken with CONFIG_PREEMPT_RT anyway as only a
subset of the functions that depend on lock type are covered by the
preprocessor #if/#else blocks.

Remove the special-case for CONFIG_PREEMPT_RT.

Signed-off-by: John Keeping <john@metanate.com>