rtl8723ds: fix sta_info alignment
Allocation function return memory that is properly aligned for most data
types, there is no need to manually align the buffer later.

On 64-bit systems this is particularly wrong as struct sta_info should
be 8-byte aligned so this calculation is guaranteed to misalign the
structure.  If the architecture requires aligned access, this leads to
an alignment fault and panic.

Signed-off-by: John Keeping <john@metanate.com>